third_party/WebKit/LayoutTests/fast/html/imports/xhr-expected.txt

Ensure XHR document doesn't load imports

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS link.import is null
PASS successfullyParsed is true

TEST COMPLETE
